% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{calc_moments}
\alias{calc_moments}
\title{Calculate moments of a numeric vector}
\usage{
calc_moments(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a named vector of descriptive statistics
}
\description{
Calculate moments of a numeric vector
}
\examples{
x <- rexp(1000, 2)
calc_moments(x)
}
