% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgets.R
\name{fit_dist_addin}
\alias{fit_dist_addin}
\title{Fit Univariate Distributions Addin}
\usage{
fit_dist_addin()
}
\description{
Interactively submit a numeric vector and choose what distributions that
you want to run fit diagnostics. Click done to have the desired distribution
code put into your cursor position.
}
