% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_empirical_continuous.R
\name{fit_empirical_continuous}
\alias{fit_empirical_continuous}
\title{Fit Continous Empirical Distribution Family}
\usage{
fit_empirical_continuous(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
list of family functions for d, p, q, r, and parameters
}
\description{
Fit Continous Empirical Distribution Family
}
\examples{
set.seed(562)
}
