% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_empirical_discrete.R
\name{fit_empirical_discrete}
\alias{fit_empirical_discrete}
\title{Fit Empirical Discrete Distribution Family}
\usage{
fit_empirical_discrete(x)
}
\arguments{
\item{x}{integer vector}
}
\value{
list of family functions for d, p, q, r, and parameters
}
\description{
Fit Empirical Discrete Distribution Family
}
\examples{
set.seed(562)
x <- rpois(100, 5)
empDis <- fit_empirical_discrete(x)
empDis$dempDis(1)
empDis$pempDis(2)
empDis$qempDis(1)
empDis$r(10000)
empDis$parameters
}
