% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{is.distfun}
\alias{is.distfun}
\title{Test if object is a distfun object}
\usage{
is.distfun(x)
}
\arguments{
\item{x}{an R object to be tested}
}
\value{
TRUE if x is a disfun object, FALSE otherwise
}
\description{
Test if object is a distfun object
}
