% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_stats}
\alias{summarize_stats}
\title{Provide Summary Stats for Inspecting Numeric Data}
\usage{
summarize_stats(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a named vector of descriptive statistics
}
\description{
Provide Summary Stats for Inspecting Numeric Data
}
\examples{
x <- rexp(1000, 2)
summarize_stats(x)
}
