% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_empirical.R
\name{fit_empirical_discrete}
\alias{fit_empirical_discrete}
\title{Fit Empirical Discrete Distribution Family}
\usage{
fit_empirical_discrete(x)
}
\arguments{
\item{x}{integer vector}
}
\value{
list of family functions for d, p, q, r, and parameters
}
\description{
Fit Empirical Discrete Distribution Family
}
