% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distfun.R
\name{gen_dist_fun}
\alias{gen_dist_fun}
\title{Generate Single Distribution Function}
\usage{
gen_dist_fun(f, parameters, ...)
}
\arguments{
\item{f}{one of distribution functions}

\item{parameters}{new parameters for distribution}

\item{...}{arguments to pass on to distribution function}
}
\value{
one of parameterized distribution functions in d, p, q, r
}
\description{
Generate Single Distribution Function
}
